[HOT-FIX] Separate Product Public with Public Venue Controller and Route
